+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the "gruvbox-material" extension will be documented in th
 ## 6.3.5
 
 - Use black as background color of remote extension icon. [#43](https://github.com/sainnhe/gruvbox-material-vscode/issues/43)
+- Add diagnostic background colors. [#44](https://github.com/sainnhe/gruvbox-material-vscode/issues/44)
 
 ## 6.3.4